refactored vfo list to separate file. cleaned up imports
